moved pos and parse rendering from sblgnt-web into commons
